requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2 